Create person_addition_module.rb and refactor for app.rb
