+++ app.rb
@@ -4,8 +4,11 @@ require_relative 'classroom'
 require_relative 'person'
 require_relative 'student'
 require_relative 'teacher'
+require_relative 'person_addition_module'
 
 class App
+  include PersonAdditionModule
+
   MENU_OPTIONS = [
     'List all books',
     'List all people',
@@ -26,60 +29,8 @@ class App
     puts
   end
 
-  def student_info
-    p 'Age:'
-    age = gets.chomp
-    p 'Name:'
-    name = gets.chomp
-    p 'Has parent permission? [Y/N]:'
-    parent_permission_input = gets.chomp.downcase
-    parent_permission = parent_permission_input == 'y' ? true : false
-    [age, name, parent_permission]
-  end
-
-  def teacher_info
-    p 'Age:'
-    age = gets.chomp
-    p 'Name:'
-    name = gets.chomp
-    p 'Specialization:'
-    specialization = gets.chomp
-    [age, name, specialization]
-  end
-
-  def create_student(classroom = @classroom)
-    age, name, parent_permission = student_info
-    new_student = Student.new(age, classroom, name: name, parent_permission: parent_permission)
-    p 'Student created successfully'
-    new_student
-  end
-
-  def create_teacher(classroom = @classroom)
-    age, name, specialization = teacher_info
-    new_teacher = Teacher.new(age, specialization, classroom, name: name)
-    p 'Teacher created successfully'
-    new_teacher
-  end
-
-  def create_person
-    p 'Do you want to create a student (1) or a teacher (2)? [input the number]: '
-    option = gets.chomp
-
-    until %w[1 2].include?(option)
-      incorrect_input
-      create_person
-    end
-
-    case option
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    end
-  end
-
   def list_of_people
-    @classroom.list_of_people
+    puts @classroom.list_of_people
   end
 
   def menu_action(option)
@@ -101,7 +52,7 @@ class App
       # Exit
       @is_active = false
     else
-    incorrect_input
+      incorrect_input
     end
   end
